Issue completed. Refactored a bunch of parts out of gameplay.rs and the subscriptions out of connector.rs
